+++ Changelog.md
@@ -12,6 +12,9 @@
   `compareFunc` and `compareFuncConstraint` rather than needing to use
   `mappend` or `<>` to manually combine them all.
 
+* Some of Criterion's command-line options are now available, though
+  none to do with saving results to file.
+
 0.1.0.0 (22 May, 2016)
 ======================